src/main/java/creationalPatterns/prototypeDesingPattern/PineTree.java
package creationalPatterns.prototypeDesingPattern;

public class PineTree extends Tree{

    private String mass;
    private String height;
    private String position;

    public PineTree(String mass, String height) {
        super();
    }

    @Override
    public Tree copy() {
        PineTree pineTreeClone = new PineTree(this.getMass(), this.getHeight());
        pineTreeClone.setPosition(this.getPosition());
        return pineTreeClone;
    }

    public String getMass() {
        return mass;
    }

    public void setMass(String mass) {
        this.mass = mass;
    }

    public String getHeight() {
        return height;
    }

    public void setHeight(String height) {
        this.height = height;
    }

    public String getPosition() {
        return position;
    }

    public void setPosition(String position) {
        this.position = position;
    }
}
